fixed the disminution problem

+++ trunk/ga-solver/src/main/java/genetics/functions/managers/CruzaManager.java
@@ -28,7 +28,7 @@ public class CruzaManager {
         BINOMIAL, MULTIPUNTO
     };
     
-    public int DEFAULT_SURVIVORS_BY_CRUZATOR_METHODS = 40;
+    public double DEFAULT_SURVIVORS_BY_CRUZATOR_METHODS = 40;
     
     private HashMap<Cruzators, ICruzator> cruzaClasses;
     static DefaultLogguer logguer = DefaultLogguer.getLogger();
@@ -62,10 +62,10 @@ public class CruzaManager {
                 int numberOfPopulation = poblacionOriginal.size();
 
                 // porcentage a pasar por este metodo
-                int percentageOfCopiesToBeGeneratedByThisMethod = getCoverageOfMethod * DEFAULT_SURVIVORS_BY_CRUZATOR_METHODS / 100;
+                double percentageOfCopiesToBeGeneratedByThisMethod = getCoverageOfMethod * DEFAULT_SURVIVORS_BY_CRUZATOR_METHODS / 100;
 
                 // total de copias de acuerdo al porcentaje anterior
-                int numberOfCopiesForThisMethod = percentageOfCopiesToBeGeneratedByThisMethod * numberOfPopulation / 100;
+                double numberOfCopiesForThisMethod = percentageOfCopiesToBeGeneratedByThisMethod * numberOfPopulation / 100;
 
                 // genero las individuos que necesito
                 while (numberOfCopiesForThisMethod > 0) {
